fix: runner not update
