+++ packages/insomnia/src/ui/routes/mock-server.tsx
@@ -21,7 +21,7 @@ import { SvgIcon } from '../components/svg-icon';
 import { OrganizationTabList } from '../components/tabs/tabList';
 import { formatMethodName } from '../components/tags/method-tag';
 import { INSOMNIA_TAB_HEIGHT } from '../constant';
-import { useInsomniaTab } from '../hooks/tab';
+import { useInsomniaTab } from '../hooks/use-insomnia-tab';
 import { MockRouteResponse, MockRouteRoute, useMockRoutePatcher } from './mock-route';
 import { useRootLoaderData } from './root';
 import type { WorkspaceLoaderData } from './workspace';
